+++ Sources/Tab.swift
@@ -13,24 +13,24 @@ public class Tab: Control {
 		clibui.uiTabAppend(self.op, name, child.control)
 	}
 
-	public func insert(name:String, before:Int, child:Control) -> Void {
-		clibui.uiTabInsertAt(self.op, name, UInt(before), child.control)
+	public func insert(name:String, before:Int32, child:Control) -> Void {
+		clibui.uiTabInsertAt(self.op, name, before, child.control)
 	}
 
-	public func delete(index:Int) -> Void {
-		clibui.uiTabDelete(self.op, UInt(index))
+	public func delete(index:Int32) -> Void {
+		clibui.uiTabDelete(self.op, index)
 	}
 
-	public func numPages() -> Int {
-		return Int(clibui.uiTabNumPages(self.op))
+	public func numPages() -> Int32 {
+		return clibui.uiTabNumPages(self.op)
 	}
 
-	public func getMargined(index:Int) -> Bool {
-		return clibui.uiTabMargined(self.op, UInt(index)) == 1
+	public func getMargined(index:Int32) -> Bool {
+		return clibui.uiTabMargined(self.op, index) == 1
 	}
 
-	public func set(margined:Bool, index:Int) {
-		clibui.uiTabSetMargined(self.op, UInt(index), margined ? 1 : 0)
+	public func set(margined:Bool, index:Int32) {
+		clibui.uiTabSetMargined(self.op, index, margined ? 1 : 0)
 	}
 
 	public func set(child:Control) -> Void {
